Form close to done
Changed file names and added author and date. Most of the form is done. The files have been linked correctly. The upper login form is done, and the main one with payment info has the states, phone, and zip done, and the default for the Country field is almost done.