Final Push - Ready for Review
Completed the form. Added the bottom part of the form. The bottom part of the form is credit card information, and includes credit card selection, credit card number, expiration dates, and the CSC field. The submit button has been added, and is an image of an arrow going right. Added CSS that changes the background color of fields when focused on and if they are either valid or invalid (does not work when the button is a radio button. Added comments